Renamed settings.py.new to settings.py
If you have to change your settings.py for local testing
do this in a local branch (test or so) and commit your changes.
Then do your edits in another branch (experimental or master or so)
and merge these edits into your test branch. This way it is easier
to notice if the settings.py changed. And it will be much better
for production use when we'll pull on our webserver(s) :)